Update configurations, improve checkpoint handling, and enhance logging
- Updated .gitignore to include pre-downloaded checkpoints while keeping the README.
- Adjusted num_beams parameter in ar_inference_articulationxl.yaml from 15 to 5.
- Added articulationxl skeleton path to tokenizer_parts_articulationxl_256.yaml and tokenizer_rignet.yaml.
- Enhanced logging in extract.py to include detailed output information.
- Improved file handling in extract.py to correctly manage absolute paths.
- Modified download.py to support additional checkpoints and improved error handling.
- Updated merge.py to ensure proper exit handling and logging.
- Changed model precision in unirig_ar.py from float32 to bfloat16.
- Altered DetokenizeOutput class in spec.py to remove frozen dataclass constraint.

Made-with: Cursor

+++ src/model/unirig_ar.py
@@ -49,8 +49,7 @@ class UniRigAR(ModelSpec):
         _d = llm.copy()
         _d['vocab_size'] = self.tokenizer.vocab_size
         llm_config = AutoConfig.from_pretrained(**_d)
-        # Force float32 precision for the model
-        llm_config.torch_dtype = torch.float32
+        llm_config.torch_dtype = torch.bfloat16
         # Force enable pre_norm
         llm_config.pre_norm = True
         self.transformer = AutoModelForCausalLM.from_config(config=llm_config)
